requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pytest-asyncio
pytest-cov
httpx